ゲームパラメータ整理
