+++ DX22Base/BossGauge.cpp
@@ -12,25 +12,27 @@
 	・2023/11/19 描画処理、ゲージ出現、消す処理追加 Tei
 	・2023/11/22 ボスゲージ表示のフェードアウト追加、パラメータ調整
 	・2023/11/27 ボス出現処理追加	Sawada
+	・2023/12/07 ゲームパラメータから一部定数移動・インクルード追加 takagi
 
 ========================================== */
 
 // =============== インクルード ===================
 #include "BossGauge.h"
+#include "Timer.h"	//STAGE_TIME用
 
 // =============== 定数定義 =======================
-
+const TPos2d<float> BOSS_GAUGE_EMPTY_POS(765.0f, 45.0f);	//ボスゲージ（空）の位置設定
+const TPos2d<float> BOSS_GAUGE_FULL_POS(765.0f, 46.5f);	//ボスゲージ（満）の位置設定
+const float BOSS_GAUGE_EMPTY_SIZE_X = 60.0f;			//ボスゲージ（空）のXの長さ設定
+const float BOSS_GAUGE_EMPTY_SIZE_Y = -60.0f;			//ボスゲージ（空）のYの長さ設定
+const float BOSS_GAUGE_FULL_SIZE_X = (6.0f / 7.0f) * BOSS_GAUGE_EMPTY_SIZE_X;			//ボスゲージ（満）のXの長さ設定
+const float BOSS_GAUGE_FULL_SIZE_Y = (6.0f / 7.0f) * BOSS_GAUGE_EMPTY_SIZE_Y;			//ボスゲージ（満）のYの長さ設定
+const float BOSS_GAUGE_FULL_POS_Y_ADJUST = BOSS_GAUGE_FULL_SIZE_X / 2;		//ボスゲージ増加時、位置表示するための調整量
+const float BOSS_GAUGE_FULL_SIZE_Y_ADJUST = BOSS_GAUGE_FULL_SIZE_Y;	//ボスゲージ増加時、サイズ計算用（計算して表示したい比率かける元々のサイズ(100.0f)）
 #if MODE_GAME_PARAMETER
 #else
 const int BOSS_GAUGE_FULL_TIME = 45 * 60;		//ボスゲージMAXになる時間(何秒出現) * 60フレーム
 const int SECOND_EMPTY_BOSS_GAUGE = 75 * 60;		//二体目のボス空ゲージ表す時間 * 60フレーム
-const TPos2d<float> BOSS_GAUGE_EMPTY_POS(765.0f, 25.0f);	//ボスゲージ（空）の位置設定
-const TPos2d<float> BOSS_GAUGE_FULL_POS(765.0f, 25.0f);	//ボスゲージ（満）の位置設定
-const float BOSS_GAUGE_EMPTY_SIZE_X = 100.0f;			//ボスゲージ（空）のXの長さ設定
-const float BOSS_GAUGE_EMPTY_SIZE_Y = -100.0f;			//ボスゲージ（空）のYの長さ設定
-const float BOSS_GAUGE_FULL_SIZE_X = 100.0f;			//ボスゲージ（満）のXの長さ設定
-const float BOSS_GAUGE_FULL_POS_Y_ADJUST = 50.0f;		//ボスゲージ増加時、位置表示するための調整量
-const float BOSS_GAUGE_FULL_SIZE_Y_ADJUST = -100.0f;		//ボスゲージ増加時、サイズ計算用（計算して表示したい比率かける元々のサイズ(-100.0f)）
 const int FADE_TIME = 5 * 60;							//ボスゲージが溜まってから消える時間
 #endif
