+++ DX22Base/CollisionOfStage.cpp
@@ -14,6 +14,7 @@
 	・2023/11/29 ヒットストップ仕様変更対応 takagi
 	・2023/11/30 ヒットストップをハンマー激突時に変更 takagi
 	・2023/12/01 ハンマーとスライムの当たり判定持にSEを再生 yamashita
+	・2023/12/07 ゲームパラメータから一部定数移動 takagi
 
 ========================================== */
 
@@ -24,6 +25,9 @@
 #include "Slime_4.h"	//赤スライム種分けよう
 #include <typeinfo.h>	//typeid使用
 
+// =============== 定数定義 =======================
+const float HAMMER_HIT_MOVE_SPEED = 1.0f;		// ハンマーに吹き飛ばされた時のスピード
+const float HIT_HAMMER_VOLUME = 0.3f;
 
 /* ========================================
    当たり判定まとめ関数
