+++ DX22Base/Explosion.cpp
@@ -25,6 +25,7 @@
 	・2023/11/21 DisplayAddTimeの中にBoooomUIの表示時間処理追加 Tei
 	・2023/12/04 爆発のエフェクトを実装 yamasita
 	・2023/12/04 爆発の仮表示3Dモデルを削除 yamasita
+	・2023/12/07 ゲームパラメータから一部定数移動 takagi
 
 ======================================== */
 
@@ -36,10 +37,10 @@
 #include "DirectWrite.h"
 
 // =============== 定数定義 =======================
+const float EXPAND_QUICK_RATE = 0.2f;			// 膨張加速割合 
 #if MODE_GAME_PARAMETER
 #else
 //const float MAX_DISPLAY_TIME = 3.0f * 60;	// 爆発持続秒数
-const float EXPAND_QUICK_RATE = 0.2f;   // 膨張加速割合 
 const int DELAY_TIME = 0.5f * 60;
 #endif
 const float EXPLODE_STANDARD_SIZE = 0.15f;
