+++ DX22Base/Scene.cpp
@@ -17,6 +17,7 @@
 	・2023/11/22 2D描画関数を追加 nieda
 	・2023/11/23 サウンドファイル読み込み関数追加 nieda
 	・2023/11/27 カメラ初期化 takagi
+	・2023/12/07 ビュー行列取得にカメラ使用 takagi
 
 ========================================== */
 
@@ -115,7 +116,7 @@ bool CScene::IsFin() const
 ========================================== = */
 void CScene::Draw2d(float posX, float posY, float h, float w, Texture* pTexture)
 {
-	DirectX::XMFLOAT4X4 mat[3];
+	DirectX::XMFLOAT4X4 mat[2];
 
 	// ワールド行列はXとYのみを考慮して作成
 	DirectX::XMMATRIX world = DirectX::XMMatrixTranslation(posX, posY, 0.0f);	// ワールド行列（必要に応じて変数を増やしたり、複数処理を記述したりする）
@@ -124,16 +125,14 @@ void CScene::Draw2d(float posX, float posY, float h, float w, Texture* pTexture)
 	// ビュー行列は2Dだとカメラの位置があまり関係ないので、単位行列を設定する
 	DirectX::XMStoreFloat4x4(&mat[1], DirectX::XMMatrixIdentity());
 
-	// プロジェクション行列には2Dとして表示するための行列を設定する
-	// この行列で2Dのスクリーンの大きさが決まる
-	DirectX::XMMATRIX proj = DirectX::XMMatrixOrthographicOffCenterLH(VIEW_LEFT, VIEW_RIGHT, VIEW_BOTTOM, VIEW_TOP, NEAR_Z, FAR_Z);	// 平衡投影行列を設定
-	DirectX::XMStoreFloat4x4(&mat[2], DirectX::XMMatrixTranspose(proj));
-
 	// スプライトの設定
 	Sprite::SetWorld(mat[0]);
 	Sprite::SetView(mat[1]);
-	Sprite::SetProjection(mat[2]);
-	Sprite::SetSize(DirectX::XMFLOAT2(h, -w));
+	if (m_pCamera)	//ヌルチェック
+	{
+		Sprite::SetProjection(m_pCamera->GetProjectionMatrix(CCamera::E_DRAW_TYPE_2D));	// 平行投影行列を設定
+	}
+	Sprite::SetSize(DirectX::XMFLOAT2(h, w));
 	Sprite::SetUVScale(DirectX::XMFLOAT2(1.0f, 1.0f));
 	Sprite::SetUVPos(DirectX::XMFLOAT2(0.0f, 0.0f));
 	Sprite::SetTexture(pTexture);
