+++ DX22Base/ScoreOHManager.cpp
@@ -10,18 +10,14 @@
 	変更履歴
 	・2023/11/18　作成 yamamoto
 	・2023/11/25　コメント訂正、使っていない関数削除 yamamoto
+	・2023/12/07 ゲームパラメータに依存していたので修正・不要箇所削除 takagi
 
 ========================================== */
 
 // =============== インクルード ===================
 #include "ScoreOHManager.h"
-#include "GameParameter.h"
-#include "Sphere.h"				//球定義ヘッダー　なんでか分からんけど付けたらましになる
-// =============== 定数定義 =======================
-#if MODE_GAME_PARAMETER
-#else
+//include "Sphere.h"				//球定義ヘッダー　なんでか分からんけど付けたらましになる
 
-#endif
 /* ========================================
 	関数：コンストラクタ
 	-------------------------------------
