+++ DX22Base/ScoreOHManager.h
@@ -10,14 +10,19 @@
    変更履歴
 	・2023/11/17　作成 yamamoto
 	・2023/11/25　使っていない関数を削除 yamamoto
+	・2023/12/07 ゲームパラメータに依存していたので修正・不要箇所削除・定数定義 takagi
+
 ========================================== */
 #ifndef __SCORE_MANAGER_H__
 #define __SCORE_MANAGER_H__
 
 // =============== インクルード ===================
 #include "ScoreOverHead.h"
-#include "GameParameter.h"
+#include "ExplosionManager.h"	//最大爆発数定義
 #include "SlimeBase.h"
+
+// =============== 定数正義 ===================
+const float SLIME_SCORE_HEIGHT = 4.0f;	//爆発時頭上スコアの表示位置
 
 // =============== クラス定義 =====================
 class CScoreOHManager
@@ -26,7 +31,6 @@ public:
 	// ===メンバ関数宣言===
 	CScoreOHManager();		//コンストラクタ
 	~CScoreOHManager();		//デストラクタ
-
 	void Update();		 		//更新関数
 	void Draw();		 		//描画関数
 	void DisplayOverheadScore(TTriType<float> pos,int score, float height);			// スコアの生成
@@ -39,9 +43,6 @@ protected:
 	CScoreOverHead* m_pScore[MAX_EXPLOSION_NUM];	//スコアの配列
 	static int m_nTotalScore[5];	//トータルスコア//複数箇所対応	//倍率かける前
 	const CCamera*	m_pCamera;	//カメラのポインタ
-
-private:
-	
 };
