+++ DX22Base/Shadow.cpp
@@ -9,13 +9,16 @@
 
 	変更履歴
 	・2023/11/27 制作 nieda
+	・2023/12/07 ゲームパラメータから一部定数移動 takagi
 
 ========================================== */
 
 // =============== インクルード ===================
 #include "Shadow.h"			// 自身のヘッダ
 #include "Sprite.h"			// 表示用ヘッダ
-#include "GameParameter.h"	// 定数定義用ヘッダ
+
+// =============== 定数定義 ===================
+const float PI = 3.14159f;					// 円周率
 
 
 /* ========================================
