+++ DX22Base/SlimeBase.cpp
@@ -26,6 +26,7 @@
 	・2023/11/28 攻撃力を追加 Sawada
 	・2023/11/29 影メモリリーク除去 takagi
 	・2023/11/30 モデルの読み込みが反転したのでradian.yが反対になるように変更 yamashita
+	・2023/12/07 ゲームパラメータから一部定数移動 takagi
 
 ========================================== */
 
@@ -36,15 +37,14 @@
 
 
 // =============== 定数定義 =======================
+const float SLIME_BASE_RADIUS = 0.5f;			// スライムの基準の大きさ
+const int	RANDOM_MOVE_SWITCH_TIME = 3 * 60;	// ランダム移動の方向切り替え
+const float SPEED_DOWN_RATIO = 0.6f;			// 吹き飛ぶ際にかかる移動速度の変化の割合    RATIO=>割合
 const float REFLECT_RATIO = 0.1f;				//スライムがスライムを吹き飛ばした際に吹き飛ばした側のスライムの移動量を変える割合
-
+const float MOVE_RESIST = 0.05f;				// 吹き飛び移動中のスライムの移動速度に毎フレームかかる減算数値
 #if MODE_GAME_PARAMETER
 #else
-const float SPEED_DOWN_RATIO = 0.7f;			// スライムが接触して吹き飛ぶ際にかかる移動速度の変化の割合	RATIO=>割合
-const float MOVE_RESIST = 0.05f;				// 吹き飛び移動中のスライムの移動速度に毎フレームかかる減算数値
 const float MOVE_DISTANCE_PLAYER = 15;			// プレイヤー追跡移動に切り替える距離
-const float SLIME_BASE_RADIUS = 0.5f;			// スライムの基準の大きさ
-const int	RANDOM_MOVE_SWITCH_TIME = 5 * 60;	// ランダム移動の方向切り替え
 #endif
 
 /* ========================================
