+++ DX22Base/Slime_1.cpp
@@ -18,6 +18,7 @@
 	・2023/11/14 SphereInfoの変更に対応 takagi
 	・2023/11/15 スライムのモデルと頂点シェーダーをmanagerから受け取るように変更 yamashita
 	・2023/11/28 影の大きさを設定する変数追加 nieda
+	・2023/12/07 ゲームパラメータから一部定数移動 takagi
 
 ========================================== */
 
@@ -26,12 +27,14 @@
 #include "GameParameter.h"		//定数定義用ヘッダー
 
 // =============== 定数定義 =======================
+const int	LEVEL1_ATTACK = 1;						// 攻撃力
 #if MODE_GAME_PARAMETER
 #else
 const float LEVEL1_SCALE = 1.0f;
 const float LEVEL1_SPEED = ENEMY_MOVE_SPEED;
-const int	LEVEL1_ATTACK = 1;	// 攻撃力
 #endif
+const float LEVEL_1_SHADOW_SCALE = 1.0f;	// スライム＿レベル１の影の大きさ
+
 /* ========================================
 	コンストラクタ関数
 	-------------------------------------
