+++ DX22Base/Slime_Flame.cpp
@@ -12,6 +12,7 @@
 	・2023/11/14 炎スライム仕様の実装 Suzumura
 	・2023/11/15 スライムのモデルと頂点シェーダーをmanagerから受け取るように変更 yamashita
 	・2023/11/28 影の大きさを設定する変数追加 nieda
+	・2023/12/07 ゲームパラメータから一部定数移動 takagi
 
 ========================================== */
 
@@ -25,8 +26,9 @@
 const float LEVEL_FLAME_SCALE = 1.0f;
 const float LEVEL_FLAME_SPEED = ENEMY_MOVE_SPEED * 0.2f;
 const int	LEVEL_FLAME_ATTACK = 1;	// 攻撃力
-
 #endif
+const float LEVEL_FLAME_SHADOW_SCALE = 2.0f;// スライム＿フレイムの影の大きさ
+
 /* ========================================
 	コンストラクタ関数
 	-------------------------------------
