+++ DX22Base/Timer.h
@@ -10,6 +10,7 @@
 	変更履歴
 	・2023/11/14 h,作成 Tei
 	・2023/11/17 開始処理を複数種類追加、タイマーストップ実装 Sawada
+	・2023/12/07 ゲームパラメータに依存していたので修正・不足インクルード解消・ゲームパラメータから一部定数移動 takagi
 
 ========================================== */
 #ifndef __TIMER_H__
@@ -18,7 +19,10 @@
 
 // =============== インクルード ===================
 #include <Texture.h>
-#include "GameParameter.h"
+#include "Pos2d.h"	//二次元座標
+
+// =============== 定数定義 ===================
+const int STAGE_TIME = 180 * 60;	//ステージ制限時間（秒*フレーム）
 
 // =============== クラス定義 =====================
 class CTimer
